requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
distlib==0.4.0
Django==5.2.8
django-bootstrap4==25.2
django-filter==25.2
djangorestframework==3.16.1
factory_boy==3.0.1
Faker==38.0.0
filelock==3.20.0
platformdirs==4.5.0
soupsieve==2.8
sqlparse==0.5.3
stevedore==5.5.0
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.35.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
